Fix file revision lookup failures caused by incorrect wiki index.
When the 'local' field is present in the PHP api output, it has
the empty string "" as its value.  This was slipping through the
conditional, which caused the bundler to think that wikitext for
all images could/should be found on commons, not the local wiki.

Also implement a fallback to commons, in case wikitext is not found
on the local wiki.

Change-Id: I46937aeff48bebb00150c1c3850069fe0b63832f